# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Students payment service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/operations/show_list": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Get Operations List",
        "operationId": "get_operations_list_operations_show_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperationResponse"
                  },
                  "type": "array",
                  "title": "Response Get Operations List Operations Show List Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/operations/new_transaction": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "New Semester Payment",
        "operationId": "new_semester_payment_operations_new_transaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/operations/add_to_group": {
      "put": {
        "tags": [
          "operations"
        ],
        "summary": "Add Student To Group",
        "operationId": "add_student_to_group_operations_add_to_group_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/operations/remove_from_group": {
      "delete": {
        "tags": [
          "operations"
        ],
        "summary": "Remove Student From Group",
        "operationId": "remove_student_from_group_operations_remove_from_group_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessfulResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GroupResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "users"
        ],
        "title": "GroupResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OperationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/OperationTypes"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "initiator": {
            "$ref": "#/components/schemas/UserOperationsResponse"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "user_id",
          "comment",
          "created_at",
          "initiator"
        ],
        "title": "OperationResponse"
      },
      "OperationTypes": {
        "type": "string",
        "enum": [
          "user",
          "group",
          "semester",
          "payment"
        ],
        "title": "OperationTypes"
      },
      "Roles": {
        "type": "string",
        "enum": [
          "student",
          "observer",
          "accountant",
          "admin"
        ],
        "title": "Roles"
      },
...
      "SuccessfulResponse": {
        "properties": {
          "success": {
            "type": "string",
            "title": "Success",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "SuccessfulResponse"
      },
...
      "TransactionCreate": {
        "properties": {
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "semester_id",
          "amount"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "semester_id",
          "amount",
          "comment",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
...
      "UserOperationsResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "surname": {
            "type": "string",
            "title": "Surname"
          },
          "patronymic": {
            "type": "string",
            "title": "Patronymic"
          }
        },
        "type": "object",
        "required": [
          "name",
          "surname",
          "patronymic"
        ],
        "title": "UserOperationsResponse"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "surname": {
            "type": "string",
            "title": "Surname"
          },
          "patronymic": {
            "type": "string",
            "title": "Patronymic"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "role": {
            "$ref": "#/components/schemas/Roles"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "surname",
          "patronymic",
          "role",
          "phone",
          "login",
          "transactions"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}